refactor(auth-service): modularize service layer
